Integrate podcast extension into the core
All podcast features are now available by default,
without needing to load the podcast extension.
This is done because this fork is meant to generate
podcasts, first and foremost.

A few remaining traces of the dc extension, forgotten
in commit ad62f1cc3d9, was also removed
from the documentation.

A mishap in the parameter type description for
itunes_complete and itunes_new_feed_url, introduced
in commit 71810044bdd, was also fixed.