Remove support for extensions
The need for extensions to the podcast format is pretty low,
since the podcasting standard probably won't change much
in the future. Instead, if you want functionality which is
missing from the core classes, you are encouraged to create
a subclass which extends the appropriate class, and overrides
methods for feed generation and initialization, and adds
getters and setters. I plan on making this much easier by
doing some refactoring; a substantial amount of boilerplate
is needed at the moment if you wish to extend it.