Rename the core classes and entry.py
FeedGenerator was renamed to Podcast, so people can think of it
as representing a single podcast. FeedGenerator implies that a
single object can be used for generating multiple, different feeds,
which is incorrect (except for creating both ATOM and RSS).

FeedEntry was renamed to Episode, to match its relationship with
Podcast. Entry is the term used in ATOM, which was abolished a few
commits ago. Its file was also renamed to item.py, to match the term
used in RSS. While it could have been renamed to episode.py, I
don't want to confuse people by having two different 'episode' pop up
when users try to figure out what to import.