Clean up on channels's pubDate vs LastBuildDate
requirements.txt is added with the pip packages you need,
to ease setting up a development environment. I have added
the pytz package at the same time, since you cannot possibly
deal with timezones without.

pubDate has a new default value, which is calculated from the
pubDate of its items. This is put to use in __main__.py.

pubDate was renamed to published, to communicate its purpose as
the date of the last time the feed was published (ie., item added).

The lastBuildDate alias for updated was removed, to better communicate
updated's purpose as the last time the content of the feed was updated.
It still has the default value of now(), which means it can be used by clients
to ensure that no new items have been published between pubDate and
now(). It is also consistent with the tag name, although the spec
says it should be "the last time the content of the channel *changed*".

Funny enough, by setting lastBuildDate to the current time, we are
making it so that the content of the channel (namely lastBuildDate)
actually was last changed at the time indicated by lastBuildDate,
if that makes any sense. A self-fulfilling prophecy!