Rename Episode to BaseEpisode, and episode_class to Episode
The idea is to make it obvious for everyone that they should use
the episode_class property. To achieve that, I'm renaming it so
it looks and behaves like a class (except it is bound to an instance
of Podcast).

But everyone would confuse Episode with Episode (rightfully so),
which is why the class called Episode is renamed to BaseEpisode,
to show its purpose as the episode class which everyone else
should extend when needed.