Remove pretty, add minimize parameter to rss methods
I think RSS feeds should be pretty by default, since it
enhances readability for people who are curious on
what the XML looks like. This doesn't increase the size
of the XML files that much if you compress with gzip.
If you want to conserve space and save transfer times,
set minimize to True.