Combine BaseEpisode.content, description and itunes:summary into summary
All the three properties content, description and itunes:summary
serve the same purpose. They are therefore combined into one
property called summary.

The very ambigious argument "type" is renamed to the boolean "html",
which indicates whether the summary is safe for HTML-parsing. Either
way, CDATA is used for both description and content.

The description is duplicated to ensure compatibility with all
clients. iTunes uses description when itunes:summary is not present,
so itunes:summary is omitted.