Rename guid to id, use enclosure url if id not set
The guid property is renamed to the more simple name 'id'.
The fact that this is a _globally unique_ identifier
is stressed in the documentation.

Furthermore, if you don't set the id but have set an
enclosure, then the enclosure's url is used as the id.
This is consistent with how most feeds do things.
A warning about this behaviour requiring the url to be
permanent is added to the documentation.

You can set id to False or an empty string to suppress
the behaviour of using the enclosure url, resulting
in an entry without a guid element.