Change from dict to Person
All attributes/methods which accept a person (be it a real
human or an organization entity), now require you to use
an instance of feedgen.person.Person. This is done to make
it obvious to users that an email and/or a name is to be used,
and help prevent previously uncaught errors if the user
misspelled a dictionary key.

This also makes all such attributes compatible with one
another, allowing you to assign one attribute's value to
another.

On the feed level, the managingEditor and itunes:author
have been combined, and support for dc:creator is added,
as the RSS Best Practices recommend. This also means the
API is changed, since all those three are set using
Podcast.author.

Note that Podcast does not support multiple
authors, even though dc:creator provides support for it.
This is planned for a future commit. The way you assign
multiple authors to an Episode has changed, though.
Now, authors are given as varargs argument, meaning
you have to unpack your list of authors.

As a sidenote, the recipe for "make test" has been
modified to ignore errors. Errors will still be
printed, but this way, all test classes will be run each time,
so you can see all the errors at once (instead of only seeing
the errors from the first failing class).