Refactor the documentation and make several improvements
Bring more to the highest level, so it's easier to navigate.

Don't have subpages – it's important to be consistent.
This also means less blah-blah index pages.

Read through some chunks of the documentation, and make
small improvements here and there.

Update the readme.md.

Take a look at setup.py to prepare for creating package tomorrow.
Also mention that dependencies must be mentioned in setup.py
inside requirements.txt.

Add page about contributing to this project.

Add waffle.io button.

Highlight the current page in the navigation.

Show the logo in the footer when in mobile mode, so mobile
readers can navigate to the front page (this isn't possible
by default!).