Add attribute for XSLT, closes #37
As a consequence, rss_file will now use rss_str internally.
This means that the rss result is loaded into memory even
when writing to files, which could mean much poorer
performance for huge feeds in particular. If this is too
much, then we might reverse this commit, but I don't think
it will be much of a problem.