Remove notify* methods, rewrite parts of the guide
The methods for notifying hubs only work with one type of hubs,
and I don't want to make a decision on behalf of the users as to
which hub they'll use (especially since the open source alternative,
Switchboard, uses some other tokens than the popular Google Appspot
and Superfeedr hub). Instead, I want the user to do that job
themselves - it is not that hard.

Update the guide to reflect that this is no longer implemented in
PodGen, and give a few pointers on how to implement it.
Warn that if you add the atom:link rel="hub" element, but don't
notify the hub of new updates, your listeners will miss out on
any new episodes.