Make requests a dependency, remove requests as parameter
Backwards incompatible!

Instead of passing requests as a parameter to all methods that
required it, it is now stored as an instance variable,
requests_session, which you can set in the Media constructor or
later. Thus, when you download PodGen, you are assured that you
can use all of its functionality without needing to install any
extras.

I don't think it's a problem to include as many libraries as we
do. Most people are likely to use requests already anyway.