Update RPM Specfile (tests, pypi, py3)
This patch updates the RPM specfile to make sure it uses modern macros
used for Python packaging. It also switches to the tarballs published on
PyPi which will include the tests in future versions. Finally, this
drops support of Python 2 which has reached end of life quite a while
ago.